+++ src/big-query-support/pull-data/getLastProcessedBlock.ts
@@ -4,7 +4,6 @@ import { BigQuery } from '@google-cloud/bigquery';
 
 import { LAST_PROCESSED_BLOCK_TABLE_ID } from '../../common';
 
-
 export const getLastProcessedBlock = async (
   bigQuery: BigQuery,
   processId: string,
