+++ src/big-query-support/push-data/insertNewMintAndNewPosition.ts
@@ -1,10 +1,9 @@
 import { BigQuery } from '@google-cloud/bigquery';
 
-import { BigQueryPositionRow } from '../../big-query-support';
-import { secondsToBqDate } from '../../big-query-support/utils';
 import { POSITIONS_TABLE_ID } from '../../common';
-import { MintOrBurnEventInfo } from '../../common/event-parsers/parseMintOrBurnEvent';
+import { MintOrBurnEventInfo } from '../../common/event-parsers';
 import { generateLpPositionRow } from '../../common/mints-and-burns/generateLpPositionRow';
+import { secondsToBqDate } from '../utils';
 
 export const insertNewMintAndNewPosition = async (
   bigQuery: BigQuery,
@@ -13,7 +12,7 @@ export const insertNewMintAndNewPosition = async (
   // console.log('Inserting new LP position following a mint...');
 
   // generate position row
-  const positionRow: BigQueryPositionRow = generateLpPositionRow(eventInfo);
+  const positionRow = generateLpPositionRow(eventInfo);
 
   const rawPositionRow = `
     \"${positionRow.marginEngineAddress}\",
