+++ src/common/event-parsers/index.ts
@@ -2,3 +2,4 @@ export * from './parseEvent';
 export * from './parseMintOrBurnEvent';
 export * from './parseSwapEvent';
 export * from './parseVAMMPriceChangeEvent';
+export * from './types';
