+++ src/lp-pnl-speed/processLpSpeedEvent/processLpSpeedEvent.ts
@@ -1,8 +1,6 @@
 import { BigQuery } from '@google-cloud/bigquery';
 
-import {
-  parseEvent,
-} from '../../common/event-parsers';
+import { parseMintOrBurnEvent, parseVAMMPriceChangeEvent } from '../../common/event-parsers';
 import { ExtendedEvent } from '../../common/types';
 import { processMintOrBurnEventLpSpeed } from './processMintOrBurnEventLpSpeed';
 import { processVAMMPriceChangeEvent } from './processVAMMPriceChangeEvent';
@@ -12,17 +10,27 @@ export const processLpSpeedEvent = async (
   event: ExtendedEvent,
   currentTick: number,
 ): Promise<number> => {
-  const eventInfo = parseEvent(event);
+  
+  switch (event.type) {
+    case 'mint': 
+    case 'burn': {
+      console.log('processing mint or burn event');
+      
+      const eventInfo = parseMintOrBurnEvent(event);
 
-  if ('tick' in eventInfo) {
-    console.log('processing vamm price change event');
-    await processVAMMPriceChangeEvent(bigQuery, eventInfo);
-    return eventInfo.tick;
-  } else if ('liquidityDelta' in eventInfo) {
-    console.log('processing mint or burn event');
-    await processMintOrBurnEventLpSpeed(bigQuery, eventInfo, currentTick);
-    return currentTick;
-  } else {
-    throw Error('Swap events are not necessary when processing lp speed events');
+      await processMintOrBurnEventLpSpeed(bigQuery, eventInfo, currentTick);
+      return currentTick;
+    }
+    case 'price_change': {
+      console.log('processing vamm price change event');
+
+      const eventInfo = parseVAMMPriceChangeEvent(event);
+
+      await processVAMMPriceChangeEvent(bigQuery, eventInfo);
+      return eventInfo.tick;
+    }
+    default: {
+      throw new Error('Swap events are not necessary when processing lp speed events');
+    }
   }
 };
