+++ src/lp-pnl/processMintEvent/index.ts
@@ -1,2 +1,2 @@
-export * from './insertNewMintAndNewPosition';
+export * from '../../big-query-support/push-data/insertNewMintAndNewPosition';
 export * from './processMintEvent';
