+++ src/mints-and-burns/processMintAndBurnEvent/processMintOrBurnEvent.ts
@@ -1,9 +1,9 @@
 import { BigQuery } from '@google-cloud/bigquery';
 
 import { pullExistingMintOrBurnRow } from '../../big-query-support';
+import { insertNewMintOrBurn } from '../../big-query-support/push-data/insertNewMintOrBurn';
 import { parseMintOrBurnEvent } from '../../common/event-parsers/parseMintOrBurnEvent';
 import { ExtendedEvent } from '../../common/types';
-import { insertNewMintOrBurn } from './insertNewMintOrBurn';
 
 export const processMintOrBurnEvent = async (
   bigQuery: BigQuery,
